streamlit_website/requirements.txt
streamlit
streamlit-option-menu
streamlit-agraph
langchain
langchain_community
langchain_neo4j
langchain_google_genai
google-generativeai
chromadb
rich
python-dotenv
tree_sitter
tree_sitter_javascript
pyyaml
neo4j
